Refactor: Split plugin_impl package from algorithm package
The purpose is to separate base classes & other support code from the
implementations. Also, to further illustrate that the plugins included
in the core of fossil aren't different from external plugins.

This hopefully should be a smooth upgrade path for those using fossil.
The rough part is that it'll forget currently trained models, but that's
kind of normal with k-means, we tend to re-train it often.