Load toots after initial GET /
It feels **a ton** faster after this change. This is also the
culmination in a series of changes in which we moved away from session
and toward a single (or few) session that roughly corresponds to a
device. Now that the model, toots, and cluster assignments are cached,
loading the index page is blazing fast!